requirements.txt
Django

djangorestframework
markdown
django-filter
django-cors-headers

django==4.2.5
django-allauth
sorl-thumbnail
pillow
django-notifications-hq